Move map components to map folder
