+++ frontend/src/components/map/MapMarker.tsx
@@ -1,19 +1,19 @@
 import { Icon } from "leaflet";
 import "leaflet/dist/leaflet.css";
 import { Marker, Popup } from "react-leaflet";
-import greenBusIcon from "../../src/images/bus-green-map-marker.svg";
-import redBusIcon from "../../src/images/bus-red-map-marker.svg";
-import whiteBusIcon from "../../src/images/bus-white-map-marker.svg";
-import greenLocationIcon from "../../src/images/location-green-map-marker.svg";
-import redLocationIcon from "../../src/images/location-red-map-marker.svg";
-import whiteLocationIcon from "../../src/images/location-white-map-marker.svg";
-import markerShadow from "../../src/images/marker-shadow.png";
-import greenPinIcon from "../../src/images/pin-green-map-marker.svg";
-import redPinIcon from "../../src/images/pin-red-map-marker.svg";
-import whitePinIcon from "../../src/images/pin-white-map-marker.svg";
-import currentPostionIcon from "../../src/images/position-map-marker.svg";
+import greenBusIcon from "../../../src/images/bus-green-map-marker.svg";
+import redBusIcon from "../../../src/images/bus-red-map-marker.svg";
+import whiteBusIcon from "../../../src/images/bus-white-map-marker.svg";
+import greenLocationIcon from "../../../src/images/location-green-map-marker.svg";
+import redLocationIcon from "../../../src/images/location-red-map-marker.svg";
+import whiteLocationIcon from "../../../src/images/location-white-map-marker.svg";
+import markerShadow from "../../../src/images/marker-shadow.png";
+import greenPinIcon from "../../../src/images/pin-green-map-marker.svg";
+import redPinIcon from "../../../src/images/pin-red-map-marker.svg";
+import whitePinIcon from "../../../src/images/pin-white-map-marker.svg";
+import currentPostionIcon from "../../../src/images/position-map-marker.svg";
 
-import { Location, LocationTypeEnum } from "../api";
+import { Location, LocationTypeEnum } from "../../api";
 
 export type MarkerProps = {
   currentLocation: Location,
